api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Denis Karabaza <denis.karabaza@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11"
fastapi = "^0.104.1"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.23"
boto3 = "^1.28.78"
alembic = "^1.12.1"
python-dotenv = "^1.0.0"
mediapipe = "^0.10.7"
ultralytics = "^8.0.207"
psycopg2-binary = "^2.9.9"
opencv-python-headless = "^4.8.1.78"
lapx = "^0.5.5"

## @NOTE: Poetry export does not work properly with these :(
#torch = [
#  { url = "https://download.pytorch.org/whl/cu118/torch-2.1.0%2Bcu118-cp311-cp311-linux_x86_64.whl", markers = "sys_platform == 'linux'" },
#  { url = "https://download.pytorch.org/whl/cu118/torch-2.1.0%2Bcu118-cp311-cp311-win_amd64.whl", markers = "sys_platform == 'win32'" }
#]
#torchvision = [
#  { url = "https://download.pytorch.org/whl/cu118/torchvision-0.16.0%2Bcu118-cp311-cp311-linux_x86_64.whl", markers = "sys_platform == 'linux'" },
#  { url = "https://download.pytorch.org/whl/cu118/torchvision-0.16.0%2Bcu118-cp311-cp311-win_amd64.whl", markers = "sys_platform == 'win32'" }
#]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"